Remember scroll position on choose6 pages
The position is remembered in a JavaScript-set cookie and restored on
page load, if the cookie exists and the chooser div is positioned to the
far left.

There is one cookie per body ID.

This feature is necessary at least for Chrome, since it doesn’t remember
the horizontal scroll position on history movements. Firefox does it
right.